Make select in DE2 editable after a selection has been made (#20)
This modifies the behavior of all select components in DE2's
configuration panel. Before it was not possible to edit a value once it
was selected. Now the cursor is automatically placed at the end of the
value and the user can use the arrow keys or backspace key to modify it.

How is this useful to users? Consider this scenario:
- I select gene ABCD1
- I decide I actually want to look at gene ABCD2
- I can now press backspace once and type a '2' in place of the '1'