move set of DE2 hardcoded "metadata slices" to the backend
A "metadata slice" is a slice ID that can be used by Data Explorer 2 as
a variable in a context or to color points.

Previously, there was a hardcoded JSON object on the frontend that
described these slices and how they can be used. Now that hardcoded data
has been moved to the backend and requested asynchronously.

That new async behavior is the first step in making context properties
sourced from Breadbox. Eventually the hardcoded data will be removed and
a request will be made to Breadbox to discover what types of metadata
are available.