update terminology used in Data Explorer 2 (#40)
This renames several properties in Data Explorer 2 in an effort to make
the terminology more consistent with Breadbox.

Mostly, I've tried to eliminate the term "entity" (and variants like
entity_type / entity_label) because its usage was often vague.

Where it was being used to denote something that could be either samples
or features, it has been replaced with "dimension" (e.g. dimension_type,
dimension_labels).

Where it was being used to indicate "the column/row you wish to use as a
slice" it has been replace with "slice" (e.g. slice_type, slice_labels).
In this way, it forms the complement to the existing terms "index" /
"index_type" / "index_labels".