Add "metadata" data type to minimal breadbox db
This data type is used by all metadata datasets, so I think it makes sense to define it in the "minimal db" requirements. Part of the reason I'm adding it here is because it would be much more complicated to ensure that the breadbox loader always adds this datatype before any metadata.