make Breadbox-compatible version of feature select component (#132)
This introduces a new DimensionSelectV2, designed to be closer aligned
to the the Breadbox API. It uses the new SliceQuery format instead of
legacy Slice IDs.

It's still very much a work and progress and has limited functionality
compared to the first version:
- there is no way to select a context
- searching for features is limited to just labels
- only limited data validation
  - For instance, there is no check to ensure a feature/sample is
    compatible with the selected dataset ID

Those issues will be addressed shortly. For now, it's only being used by
the proof-of-concept ContextBuilderV2.