improve Elara feature selection (#143)
This brings the new feature selection component (DimensionSelectV2)
closer to the behavior of the original. It leverages some new Breadbox
APIs to implement auto-selection and tooltips to help you figure out
which features are compatible with which datasets.

It now works well enough to be suitable for many use cases but it does
not yet allow you to select a context or get details about the selected
data version.