comment out usage of slow query params
The /breadbox/types/dimensions/gene/identifiers endpoint has a query
parameter called ?show_only_dimensions_in_datasets that you can use to
ensure that all of the IDs returned are ones that are  represented in a
dataset somewhere.

The reasoning behind this is that there's no point in showing the users
options that are a dead end. While that's a nicety, it's not crucial
because we already filter the options based on data type. That means
it's impossible to make an invalid selection anyway.