Reposition Inferred Molecular Subtype in Context Manager
Asana: https://app.asana.com/0/1165651979405609/1209804368349686/f

This modifies the UI to treat one-hot encoded matrices more gracefully.
Instead of exposing the details of the encoding to the user, the columns
are displayed as pseudo-categories they can choose from.

For now, only the Molecular Subtypes dataset behaves this way. I've done
some experiments that suggest this could be a good approach for handling
all of our one-hot encoded matrices.