add flexible utility for caching API requests (#310)
Say you want this request to be cached:
  const datasets = await breadboxAPI.getDatasets();

Now you can do this:
  import { breadboxAPI, cached } from "@depmap/api";
  const datasets = await cached(breadboxAPI).getDatasets();

This is only an in-memory cache and will be cleared when the page
reloads. Also note that responses are cached indefinitely. That's how
Data Explorer 2 has worked for a long time and it's never become a
problem. But I may turn this into a LRU or provide options for
invalidating the cache in the future.