fix bug where Cell Line Selector could get stuck in loading state
This would only happen when you reached it from Featured Tools section
on the homepage. That opened the old (and rather broken) version of Cell
Line Selector.

The solution is to modify the behavior so it takes you Context Manager
instead. This is better because it launches in a mode that explains how
Context Manager and Cell Line Selector have merged into one tool.