add new @depmap/react-table package
This is a wrapper around V8 of @tanstack/react-table. That library is
"headless" which means you have to implement the UI elements yourself,
which I've done here.

This component is still at an experimental stage. Eventually it will
serve as the basis for a new version of wide-table (which is still using
V7 of @tanstack/react-table).