update UI to no longer use labels from /datasets/dimension/data/ (#383)
We want to rework this endpoint to only return a set of `ids` and
`values`. The responses can be quite large so eliminating `labels`
will help shrink them (at least a little bit).

`labels` can be fetched from /types/dimensions/{name}/identifiers
instead.

This commit updates the UI to do exactly that so it will be safe to
make the change on the backend.