prefer other endpoints over /datasets/dimension/data/ in DE2 UI (#384)
That endpoint's response format includes not just ids and values but
also the labels for each feature/sample. In the case of the dimension
type "transcript" (which has around 160K features), that can cause the
request to fail.

More context:
https://app.asana.com/1/9513920295503/project/1165651979405609/task/1211048526298770

These two endpoints can be used to achieve similar results:
/datasets/matrix/{dataset_id}
/datasets/tabular/{dataset_id}

The idea is to replace any request that uses a slice query with an
equivalent subset operation (i.e. a slice can be thought of as subset
based on exactly one sample id, feature id, or column name).

While being less ergonomic, those endpoints respond with only values
indexed by ids, omitting labels completely, so they're preferable for
requesting large slices.

This commit reworks a few operations in the Data Explorer UI to work
that way.