fix(breadbox): support dataset given IDs containing slashes
For the most part part a `given_id` can be used wherever a regular ID
would appear. However, this doesn't work quite right in URLs. `given_id`
values are arbitrary strings, not simple IDs, and may include slashes.
This breaks FastAPI/Starlette routing, since URLs are decoded before
route matching. A slash inside an ID is treated as a path delimiter
even if the value was URL-encoded 🫤. This is a known issue:
https://github.com/fastapi/fastapi/issues/791
https://github.com/fastapi/fastapi/discussions/7328

For example, this given_id works just fine:
https://cds.team/depmap/breadbox/datasets/ssgsea

While this one doesn't (without this fix):
https://cds.team/depmap/breadbox/datasets/mirna-573f.3%2FCCLE_miRNA_MIMAT

Workaround: declare the route parameter with `:path` like this:
@router.get("/{dataset_id:path}")

That allows the route to match properly and the URL-encoded ID to be
processed.

One caveat: `:path` consumes the remainder of the URL. Trying to use it
in a nested route like "/types/dimensions/{name:path}/identifiers" will
not work. Fortunately, all our other routes have the variable portion at
the end so this workaround is acceptable.