Changes to custom analysis to reduce memory utilization (#461)
A large rework of a fair amout of the custom analysis code in order to get correlation calculation to operate in a streaming fashion.

Memory footprint dropped from 1,206 MB to 325 MB after changes.

Also in the course of making these changes, discovered the p-value correction was being applied at the wrong step in the process, resulting in less of a correction then should have been applied. These changes also include a fix for that mistake.

(Also the progress % is now a real estimate)

note: the run_lm code path still does not stream