Commented out GET /metadata endpoint
Test broke and the behavior of this endpoint seems a little confusing/flaky in general. I'm not 100% sure that it makes sense as-is.

Comments suggest it was only created for Elara, so commented out code instead of changing)

(Leaving it commented for the moment to make it easier to trace back in git blame why this was removed)